view user account, get all user account, delete user account, all the remained functions for student.

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/EmptySlotRepository.java
@@ -8,9 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EmptySlotRepository extends JpaRepository<EmptySlot, Integer> {
+public interface EmptySlotRepository extends JpaRepository<EmptySlot, Long> {
 
     @Query("select e from EmptySlot e where e.student.userId = :userId")
     List<EmptySlot> findEmptySlotsByUser_UserId(Long userId);
+
 
 }
